+++ lib/dart_web_scraper/parsers/string_between_parser.dart
@@ -2,11 +2,11 @@ import 'package:dart_web_scraper/common/utils/data_extraction.dart';
 import 'package:html/dom.dart';
 import 'package:dart_web_scraper/dart_web_scraper.dart';
 
-Data? stringBetweenParser(
-  Parser parser,
-  Data parentData,
-  bool debug,
-) {
+Data? stringBetweenParser({
+  required Parser parser,
+  required Data parentData,
+  required bool debug,
+}) {
   printLog("----------------------------------", debug, color: LogColor.yellow);
   printLog(
     "ID: ${parser.id} Parser: String Between",